+++ lib/mac/say/voices_attributes.rb
@@ -1,0 +1,510 @@
+# Additional voice attributes mixed to the original Hash of
+# parsed voices (to provide additional context when looking for a voice)
+#
+# @note Unfortunately information in system *.plist files
+#   e.g. /System/Library/Speech/Voices/<VoiceName>.SpeechVoice/Contents/Info.plist"
+#   is inconsistent, fragmentary and sometimes incorrect, but:
+#
+#   WARNING: all the attributes values provided here are highly subjective!!!
+ADDITIONAL_VOICE_ATTRIBUTES = {
+    _unknown_voice: {
+      gender: nil,
+      joke: nil,
+      quality: nil
+    },
+    agnes: {
+      gender: :female,
+      joke: false,
+      quality: :low
+    },
+    albert: {
+      gender: :male,
+      joke: true,
+      quality: :medium
+    },
+    alex: {
+      gender: :male,
+      joke: false,
+      quality: :medium
+    },
+    alice: {
+      gender: :female,
+      joke: false,
+      quality: :high
+    },
+    allison: {
+      gender: :female,
+      joke: false,
+      quality: :high
+    },
+    alva: {
+      gender: :female,
+      joke: false,
+      quality: :low
+    },
+    amelie: {
+      gender: :female,
+      joke: false,
+      quality: :medium
+    },
+    angelica: {
+      gender: :female,
+      joke: false,
+      quality: :low
+    },
+    anna: {
+      gender: :female,
+      joke: false,
+      quality: :medium
+    },
+    audrey: {
+      gender: :female,
+      joke: false,
+      quality: :medium
+    },
+    aurelie: {
+      gender: :female,
+      joke: false,
+      quality: :medium
+    },
+    ava: {
+      gender: :female,
+      joke: false,
+      quality: :high
+    },
+    bahh: {
+      gender: :male,
+      joke: true,
+      quality: :low
+    },
+    bells: {
+      gender: :neutral,
+      joke: true,
+      quality: :medium
+    },
+    boing: {
+      gender: :neutral,
+      joke: true,
+      quality: :medium
+    },
+    bruce: {
+      gender: :male,
+      joke: false,
+      quality: :low
+    },
+    bubbles: {
+      gender: :male,
+      joke: true,
+      quality: :medium
+    },
+    carlos: {
+      gender: :male,
+      joke: false,
+      quality: :high
+    },
+    carmit: {
+      gender: :female,
+      joke: false,
+      quality: :medium
+    },
+    catarina: {
+      gender: :female,
+      joke: false,
+      quality: :low
+    },
+    cellos: {
+      gender: :male,
+      joke: true,
+      quality: :medium
+    },
+    cem: {
+      gender: :male,
+      joke: false,
+      quality: :medium
+    },
+    chantal: {
+      gender: :female,
+      joke: false,
+      quality: :medium
+    },
+    claire: {
+      gender: :female,
+      joke: false,
+      quality: :medium
+    },
+    damayanti: {
+      gender: :female,
+      joke: false,
+      quality: :medium
+    },
+    daniel: {
+      gender: :male,
+      joke: false,
+      quality: :high
+    },
+    deranged: {
+      gender: :neutral,
+      joke: true,
+      quality: :medium
+    },
+    diego: {
+      gender: :male,
+      joke: false,
+      quality: :low
+    },
+    ellen: {
+      gender: :female,
+      joke: false,
+      quality: :low
+    },
+    ewa: {
+      gender: :female,
+      joke: false,
+      quality: :medium
+    },
+    federica: {
+      gender: :female,
+      joke: false,
+      quality: :medium
+    },
+    felipe: {
+      gender: :male,
+      joke: false,
+      quality: :medium
+    },
+    fiona: {
+      gender: :female,
+      joke: false,
+      quality: :high
+    },
+    fred: {
+      gender: :male,
+      joke: true,
+      quality: :low
+    },
+    henrik: {
+      gender: :male,
+      joke: false,
+      quality: :high
+    },
+    hysterical: {
+      gender: :male,
+      joke: true,
+      quality: :medium
+    },
+    ioana: {
+      gender: :female,
+      joke: false,
+      quality: :low
+    },
+    iveta: {
+      gender: :female,
+      joke: false,
+      quality: :low
+    },
+    joana: {
+      gender: :female,
+      joke: false,
+      quality: :low
+    },
+    jorge: {
+      gender: :male,
+      joke: false,
+      quality: :low
+    },
+    juan: {
+      gender: :male,
+      joke: false,
+      quality: :medium
+    },
+    junior: {
+      gender: :male,
+      joke: true,
+      quality: :low
+    },
+    kanya: {
+      gender: :female,
+      joke: false,
+      quality: :medium
+    },
+    karen: {
+      gender: :female,
+      joke: false,
+      quality: :high
+    },
+    kate: {
+      gender: :female,
+      joke: false,
+      quality: :high
+    },
+    kathy: {
+      gender: :female,
+      joke: true,
+      quality: :low
+    },
+    katya: {
+      gender: :female,
+      joke: false,
+      quality: :medium
+    },
+    klara: {
+      gender: :female,
+      joke: false,
+      quality: :low
+    },
+    kyoko: {
+      gender: :female,
+      joke: false,
+      quality: :low
+    },
+    laura: {
+      gender: :female,
+      joke: false,
+      quality: :low
+    },
+    lee: {
+      gender: :male,
+      joke: false,
+      quality: :high
+    },
+    lekha: {
+      gender: :female,
+      joke: false,
+      quality: :low
+    },
+    luca: {
+      gender: :male,
+      joke: false,
+      quality: :high
+    },
+    luciana: {
+      gender: :female,
+      joke: false,
+      quality: :low
+    },
+    maged: {
+      gender: :male,
+      joke: false,
+      quality: :low
+    },
+    magnus: {
+      gender: :male,
+      joke: false,
+      quality: :medium
+    },
+    mariska: {
+      gender: :female,
+      joke: false,
+      quality: :medium
+    },
+    markus: {
+      gender: :male,
+      joke: false,
+      quality: :low
+    },
+    :'mei-jia' => {
+      gender: :female,
+      joke: false,
+      quality: :medium
+    },
+    melina: {
+      gender: :female,
+      joke: false,
+      quality: :low
+    },
+    milena: {
+      gender: :female,
+      joke: false,
+      quality: :medium
+    },
+    moira: {
+      gender: :female,
+      joke: false,
+      quality: :medium
+    },
+    monica: {
+      gender: :female,
+      joke: false,
+      quality: :medium
+    },
+    nicolas: {
+      gender: :male,
+      joke: false,
+      quality: :high
+    },
+    nikos: {
+      gender: :male,
+      joke: false,
+      quality: :medium
+    },
+    nora: {
+      gender: :female,
+      joke: false,
+      quality: :low
+    },
+    oliver: {
+      gender: :male,
+      joke: false,
+      quality: :low
+    },
+    oskar: {
+      gender: :male,
+      joke: false,
+      quality: :medium
+    },
+    otoya: {
+      gender: :male,
+      joke: false,
+      quality: :medium
+    },
+    paola: {
+      gender: :female,
+      joke: false,
+      quality: :high
+    },
+    paulina: {
+      gender: :female,
+      joke: false,
+      quality: :low
+    },
+    petra: {
+      gender: :female,
+      joke: false,
+      quality: :medium
+    },
+    princess: {
+      gender: :female,
+      joke: true,
+      quality: :low
+    },
+    ralph: {
+      gender: :male,
+      joke: true,
+      quality: :low
+    },
+    samantha: {
+      gender: :female,
+      joke: false,
+      quality: :high
+    },
+    sara: {
+      gender: :female,
+      joke: false,
+      quality: :medium
+    },
+    satu: {
+      gender: :female,
+      joke: false,
+      quality: :medium
+    },
+    serena: {
+      gender: :female,
+      joke: false,
+      quality: :medium
+    },
+    :'sin-ji' => {
+      gender: :female,
+      joke: false,
+      quality: :medium
+    },
+    soledad: {
+      gender: :female,
+      joke: false,
+      quality: :medium
+    },
+    susan: {
+      gender: :female,
+      joke: false,
+      quality: :medium
+    },
+    tarik: {
+      gender: :male,
+      joke: false,
+      quality: :low
+    },
+    tessa: {
+      gender: :female,
+      joke: false,
+      quality: :high
+    },
+    thomas: {
+      gender: :male,
+      joke: false,
+      quality: :high
+    },
+    :'ting-ting' => {
+      gender: :female,
+      joke: false,
+      quality: :high
+    },
+    tom: {
+      gender: :male,
+      joke: false,
+      quality: :high
+    },
+    trinoids: {
+      gender: :neutral,
+      joke: true,
+      quality: :medium
+    },
+    veena: {
+      gender: :female,
+      joke: false,
+      quality: :high
+    },
+    vicki: {
+      gender: :female,
+      joke: false,
+      quality: :low
+    },
+    victoria: {
+      gender: :female,
+      joke: false,
+      quality: :low
+    },
+    whisper: {
+      gender: :male,
+      joke: true,
+      quality: :medium
+    },
+    xander: {
+      gender: :male,
+      joke: false,
+      quality: :high
+    },
+    yannick: {
+      gender: :male,
+      joke: false,
+      quality: :medium
+    },
+    yelda: {
+      gender: :female,
+      joke: false,
+      quality: :low
+    },
+    yuna: {
+      gender: :female,
+      joke: false,
+      quality: :medium
+    },
+    yuri: {
+      gender: :male,
+      joke: false,
+      quality: :low
+    },
+    zarvox: {
+      gender: :neutral,
+      joke: true,
+      quality: :medium
+    },
+    zosia: {
+      gender: :female,
+      joke: false,
+      quality: :low
+    },
+    zuzana: {
+      gender: :female,
+      joke: false,
+      quality: :high
+    }
+}.freeze